Hummer Images
I got a few hummer images that are free to use without worrying about
people claiming ownership, and then I also got a few images that aren’t
explicitly listed as free to use. We should be fine with those because
if someone makes a complaint about us using their image we should just
be given a warning to take it off the site.